requirements.txt
pytesseract
Pillow
beautifulsoup4
requests
pyliblouis
lxml